--- a/BenchmarkDotNet.Artifacts/results/DSInternals.Json.Benchmarks.JsonDecodeBench-report.html
+++ b/BenchmarkDotNet.Artifacts/results/DSInternals.Json.Benchmarks.JsonDecodeBench-report.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html lang='en'>
+<head>
+<meta charset='utf-8' />
+<title>DSInternals.Json.Benchmarks.JsonDecodeBench-20250817-075533</title>
+
+<style type="text/css">
+	table { border-collapse: collapse; display: block; width: 100%; overflow: auto; }
+	td, th { padding: 6px 13px; border: 1px solid #ddd; text-align: right; }
+	tr { background-color: #fff; border-top: 1px solid #ccc; }
+	tr:nth-child(even) { background: #f8f8f8; }
+</style>
+</head>
+<body>
+<pre><code>
+BenchmarkDotNet v0.15.2, Windows 11 (10.0.26100.4652/24H2/2024Update/HudsonValley)
+Unknown processor
+.NET SDK 9.0.304
+  [Host]     : .NET 8.0.19 (8.0.1925.36514), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
+  DefaultJob : .NET 8.0.19 (8.0.1925.36514), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
+</code></pre>
+<pre><code></code></pre>
+
+<table>
+<thead><tr><th>Method     </th><th>Mean </th><th>Error</th><th>StdDev</th><th>Median</th><th>Gen0</th><th>Gen1</th><th>Gen2</th><th>Allocated</th>
+</tr>
+</thead><tbody><tr><td>STJ_String_Small</td><td>279.8 ns</td><td>4.15 ns</td><td>3.88 ns</td><td>279.0 ns</td><td>0.0448</td><td>-</td><td>-</td><td>752 B</td>
+</tr><tr><td>STJ_Bytes_Small</td><td>342.9 ns</td><td>6.37 ns</td><td>5.96 ns</td><td>342.5 ns</td><td>0.1063</td><td>0.0005</td><td>-</td><td>1784 B</td>
+</tr><tr><td>STJ_String_Medium</td><td>631.9 ns</td><td>28.32 ns</td><td>83.50 ns</td><td>649.3 ns</td><td>0.2584</td><td>0.0038</td><td>-</td><td>4336 B</td>
+</tr><tr><td>STJ_Bytes_Medium</td><td>777.3 ns</td><td>16.94 ns</td><td>48.33 ns</td><td>763.0 ns</td><td>0.7496</td><td>0.0334</td><td>-</td><td>12536 B</td>
+</tr><tr><td>STJ_String_Large</td><td>4,043.0 ns</td><td>66.74 ns</td><td>52.11 ns</td><td>4,041.0 ns</td><td>3.9215</td><td>0.7782</td><td>-</td><td>65776 B</td>
+</tr><tr><td>STJ_Bytes_Large</td><td>36,013.7 ns</td><td>743.78 ns</td><td>2,193.05 ns</td><td>35,435.2 ns</td><td>41.6565</td><td>41.6565</td><td>41.6565</td><td>196884 B</td>
+</tr></tbody></table>
+</body>
+</html>