--- a/BenchmarkDotNet.Artifacts/results/DSInternals.Json.Benchmarks.LapsBench-report.html
+++ b/BenchmarkDotNet.Artifacts/results/DSInternals.Json.Benchmarks.LapsBench-report.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html lang='en'>
+<head>
+<meta charset='utf-8' />
+<title>DSInternals.Json.Benchmarks.LapsBench-20250817-080046</title>
+
+<style type="text/css">
+	table { border-collapse: collapse; display: block; width: 100%; overflow: auto; }
+	td, th { padding: 6px 13px; border: 1px solid #ddd; text-align: right; }
+	tr { background-color: #fff; border-top: 1px solid #ccc; }
+	tr:nth-child(even) { background: #f8f8f8; }
+</style>
+</head>
+<body>
+<pre><code>
+BenchmarkDotNet v0.15.2, Windows 11 (10.0.26100.4652/24H2/2024Update/HudsonValley)
+Unknown processor
+.NET SDK 9.0.304
+  [Host]     : .NET 8.0.19 (8.0.1925.36514), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
+  DefaultJob : .NET 8.0.19 (8.0.1925.36514), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
+</code></pre>
+<pre><code></code></pre>
+
+<table>
+<thead><tr><th>Method    </th><th>Mean</th><th>Error</th><th>StdDev</th><th>Gen0</th><th>Allocated</th>
+</tr>
+</thead><tbody><tr><td>Laps_String</td><td>246.7 ns</td><td>6.61 ns</td><td>18.96 ns</td><td>0.0124</td><td>208 B</td>
+</tr><tr><td>Laps_Utf8_Bytes</td><td>259.3 ns</td><td>6.84 ns</td><td>20.16 ns</td><td>0.0219</td><td>368 B</td>
+</tr><tr><td>Laps_Utf16_Bytes</td><td>281.1 ns</td><td>8.17 ns</td><td>24.08 ns</td><td>0.0219</td><td>368 B</td>
+</tr></tbody></table>
+</body>
+</html>