--- a/Src/DSInternals.Json.Benchmarks/Program.cs
+++ b/Src/DSInternals.Json.Benchmarks/Program.cs
@@ -1,0 +1,10 @@
+﻿using BenchmarkDotNet.Running;
+
+namespace DSInternals.Json.Benchmarks
+{
+    public class Program
+    {
+        public static void Main(string[] args) =>
+            BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
+    }
+}