@MichaelGrafnetter 2025-08-17
Hi, thanks. Please send any changes to the parser in a standalone PR, as I do not want to get the benchmarks and results into the repo.

My question still stays unanswered though: What impact would migration from byte[] to Memory<> have on reading 100K or 1M computers from an AD DB? Would the decreased number of allocations be even noticeable? Similarly with parsing secrets and supplemental credentials.