@PrzemyslawKlys 2025-08-17
I submitted another PR which seems to have better results the this one. I will clean PR from the code you don't need once (and if) you decide to have it. I wanted to make sure you have a full picture from tests and not just blindly trust in results. 

The summary is for the new version, not this one, but it works differently, and required a test change to accommodate for it.

## Impact of Migrating from `byte[]` to `Span<>/Memory<>`

### 🔑 What changes
- Operate on **slices** of the original buffer instead of allocating new `byte[]`.
- Use APIs that accept `ReadOnlySpan<byte>` (JSON, crypto, ASN.1).
- Allocate only when materializing final strings or arrays.

### 📊 Impact at scale
Even modest per-record savings add up:

| Per-record saved | 100K records | 1M records |
|------------------|-------------:|-----------:|
| **~1 KB**        | ~100 MB less | ~1 GB less |
| **~8 KB**        | ~0.8 GB less | ~8 GB less |
| **~128 KB**      | ~12 GB less  | ~122 GB less |

- Benchmarks already show **2/3 fewer allocations** and up to **10× faster** for large UTF-8 JSON.
- Less GC pressure → fewer pauses, smoother throughput.

### 🧩 Where it helps most
- **AD exports / KeyCredential parsing**: slice BLOBs, avoid `ReadBytes`.
- **LAPS JSON / device keys**: deserialize directly from `ReadOnlySpan<byte>`.
- **Secrets & supplemental credentials**: span-based crypto imports (no `.ToArray()`).

At 100K–1M objects:
- Allocation reduction is **absolutely noticeable** (hundreds of MB to >100 GB depending on payload size).
- For copy-heavy JSON paths, throughput can improve **multiplicatively**.
- For crypto-heavy paths, wall-clock wins are smaller but GC pressure still drops.